# cyber-chat-with-ORM/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/comments": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The Comment has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Comment"
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "comments"
        ]
      },
      "get": {
        "operationId": "findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Users retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommentResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "summary": "List all comments in a thread related to the logged user",
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/{id}": {
      "get": {
        "operationId": "findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "comment with the specified ID was not found."
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "comments"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "comments"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "comments"
        ]
      }
    },
    "/comments/thread/{threadId}": {
      "get": {
        "operationId": "findAllByThreadId",
        "parameters": [
          {
            "name": "threadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "token": []
          }
        ],
        "tags": [
          "comments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cyber Chat API",
    "description": "Threads, user, and comments",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "threadId": {
            "type": "string",
            "example": "111-111-1111-111",
            "description": "Unique UUID threadid"
          },
          "body": {
            "type": "string",
            "example": " Swagger is a good tool to use",
            "description": "comment text"
          }
        },
        "required": [
          "threadId",
          "body"
        ]
      },
...
      "Comment": {
        "type": "object",
        "properties": {}
      },
      "CommentResponseDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateCommentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}